+++ python_modules/automation/automation/helm/schema/subschema/kubernetes.py
@@ -87,3 +87,8 @@ class StartupProbe(BaseModel):
 class SecretRef(BaseModel):
     class Config:
         schema_extra = {"$ref": create_definition_ref("io.k8s.api.core.v1.LocalObjectReference")}
+
+
+class SecretEnvSource(BaseModel):
+    class Config:
+        schema_extra = {"$ref": create_definition_ref("io.k8s.api.core.v1.SecretEnvSource")}