+++ python_modules/automation/automation/helm/schema/values.py
@@ -19,3 +19,4 @@ class HelmValues(BaseModel):
     ingress: subschema.Ingress
     imagePullSecrets: List[subschema.kubernetes.SecretRef]
     computeLogManager: subschema.ComputeLogManager
+    scheduler: subschema.Scheduler